pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm[toml]>=6.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project]
name = "pyneon"
version = "0.1.0"
description = "A Python library for reading, preprocessing and exporting Pupil Labs Neon data"
readme = "README.md"
authors = [
    { name = "Qian Chu", email = "qianchu99@gmail.com" },
    { name = "Jan-Gabriel Hartel", email = "jangabrielhartel@gmail.com" },
    { name = "Alex Lepauvre", email = "alex.lepauvre@gmail.com" },
    { name = "Lucia Melloni"}
]
requires-python = ">=3.10"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
]
dependencies = [
    "pandas",
    "matplotlib",
    "scipy",
    "opencv-python>=4.7",
    "joblib",
    "typeguard",
    "requests",
    "tqdm",
]
license = "MIT"

[project.optional-dependencies]
dev = ["pytest"]
doc = [
    "seedir",
    "ruff",
    "isort",
    "jupyter",
    "sphinx",
    "numpydoc",
    "pydata-sphinx-theme",
    "nbsphinx",
    "myst-parser",
]
full = ["pyneon[dev,doc]"]

[project.urls]
Homepage = "https://ncc-brain.github.io/PyNeon/"
Issues = "https://github.com/ncc-brain/PyNeon/issues"

[tool.pytest.ini_options]
addopts = "--code-highlight=no"

